// liubai-backends/liubai-laf/cloud-functions/common-types.ts
export interface LiuRqReturn<T = Record<string, any>> {
  code: string
  errMsg?: string
  showMsg?: string
  data?: T
}
// ...
interface BaseTable {
  _id: string
  insertedStamp: number
  updatedStamp: number
}
// ...
export interface Config_WeChat_GZH {

  // about access token
  access_token?: string
  expires_in?: number      // how many seconds left before it expires, which is only for access_token
  lastGetStamp?: number    // Timestamp of the last time access_token was obtained

  // about jsapi_ticket for js-sdk
  jsapi_ticket?: string
}

export interface Config_WeChat_MINI {
  access_token?: string
  expires_in?: number 
  lastGetStamp?: number
}

export interface Config_WeCom_Qynb {
  access_token?: string
  expires_in?: number
  lastGetStamp?: number
}

export interface LiuWxpayCert {
  serial_no: string
  effective_time: string
  expire_time: string
  cert_pem: string
}

export interface Table_Config extends BaseTable {

  // RSA Key Pair in PEM format
  publicKey: string
  privateKey: string

  // AES-GCM Key in base64 format
  aesKey?: string
  // AES-GCM IV in base64 format
  aesIV?: string

  // wechat subscription
  wechat_gzh?: Config_WeChat_GZH

  // wechat miniprogram
  wechat_mini?: Config_WeChat_MINI

  // wecom config for company internal development
  // qynb means 企业内部（开发）
  wecom_qynb?: Config_WeCom_Qynb

  wxpay_certs?: LiuWxpayCert[]

}
// ...
  export interface Param {
    operateType: "get-upload-token"
    purpose?: "avatar" | "coupon-upload" | "coupon-tmp"
  }
// ...
export namespace ServicePolyAPI {
  
  export interface Param {
    operateType: "get-wxjssdk-config"
    url: string
  }

  export interface Res_GetWxjssdkConfig {
    operateType: "get-wxjssdk-config"
    appId: string
    timestamp: number
    nonceStr: string
    signature: string
  }
}